write_basis_data_to_csv + fixtures moved
