+++ source/databricks/package/calculator_job.py
@@ -14,7 +14,7 @@
 
 from datetime import datetime
 import sys
-from pyspark.sql import SparkSession
+from pyspark.sql import SparkSession, DataFrame
 from pyspark.sql.functions import col
 import ast
 
@@ -103,6 +103,19 @@ def _get_valid_args_or_throw():
     return args
 
 
+def write_basis_data_to_csv(
+    data_df: DataFrame, process_results_path: str, batch_id: str, resolution_type: str
+):
+    (
+        data_df.withColumnRenamed("GridAreaCode", "grid_area")
+        .repartition("grid_area")
+        .write.mode("overwrite")
+        .partitionBy("grid_area")
+        .option("header", True)
+        .csv(f"{process_results_path}/basis-data/batch_id={batch_id}/{resolution_type}")
+    )
+
+
 def start(spark: SparkSession, args):
     # Merge schema is expensive according to the Spark documentation.
     # Might be a candidate for future performance optimization initiatives.
@@ -116,7 +129,7 @@ def start(spark: SparkSession, args):
         args.time_series_points_path
     )
 
-    (result_df, timeseries_basis_data) = calculate_balance_fixing_total_production(
+    (result_df, timeseries_basis_data_df) = calculate_balance_fixing_total_production(
         raw_integration_events_df,
         raw_time_series_points_df,
         args.batch_id,
@@ -129,28 +142,22 @@ def start(spark: SparkSession, args):
 
     debug("raw_timeseries", raw_time_series_points_df)
 
-    (timeseries_quarter_df, timeseries_hour_df) = timeseries_basis_data
+    (timeseries_quarter_df, timeseries_hour_df) = timeseries_basis_data_df
     debug("timeseries basis data df_hour", timeseries_hour_df)
     debug("timeseries basis data df_quarter", timeseries_quarter_df)
-    (
-        timeseries_quarter_df.withColumnRenamed("GridAreaCode", "grid_area")
-        .repartition("grid_area")
-        .write.mode("overwrite")
-        .partitionBy("grid_area")
-        .option("header", True)
-        .csv(
-            f"{args.process_results_path}/basis-data/batch_id={args.batch_id}/time-series-quarter"
-        )
+
+    write_basis_data_to_csv(
+        timeseries_quarter_df,
+        args.process_results_path,
+        args.batch_id,
+        "time-series-quarter",
     )
-    (
-        timeseries_hour_df.withColumnRenamed("GridAreaCode", "grid_area")
-        .repartition("grid_area")
-        .write.mode("overwrite")
-        .partitionBy("grid_area")
-        .option("header", True)
-        .csv(
-            f"{args.process_results_path}/basis-data/batch_id={args.batch_id}/time-series-hour"
-        )
+
+    write_basis_data_to_csv(
+        timeseries_hour_df,
+        args.process_results_path,
+        args.batch_id,
+        "time-series-hour",
     )
 
     # First repartition to co-locate all rows for a grid area on a single executor.
